Add newer versions of Eclipse to the profile list.
Had some trouble installing the connector on Kepler because of
dependencies (like json 1.0.0 and some of the logging parts)
Adding these, making Indigo non-default, and changing the
platform-version-name property to "kepler" produced a build
that was happy to install.

Perhaps combining this with newer Orbit drops for the profiles
on newer versions may also be sensible.