Include org.json bundle in feature/p2 repo
This bundle is located in Orbit repo, so there is a chance
that user does not have it in p2 list inside his Eclipse.
Bundling org.json fixes this potential problems.

In all Orbit repos from Juno to Mars, the org.json bundle 
is published with the same version 1.0.0.v201011060100,
so there is minimal risk that update will apear.